+++ frontend/src/components/modules/about/about-cta.tsx
@@ -1,0 +1,78 @@
+"use client"
+
+import Link from "next/link"
+import { motion } from "framer-motion"
+import { ArrowRight } from "lucide-react"
+import { Button } from "@/components/ui/button"
+
+export default function AboutCTA() {
+  return (
+    <section className="relative overflow-hidden bg-black py-20 text-white">
+      <div className="absolute inset-0 z-0">
+        <div className="absolute inset-0 bg-[url('/placeholder.svg?height=600&width=1200')] bg-cover bg-center opacity-10 mix-blend-overlay"></div>
+
+        {/* Animated particles */}
+        <div className="absolute inset-0">
+          {[...Array(20)].map((_, i) => (
+            <motion.div
+              key={i}
+              className="absolute h-2 w-2 rounded-full bg-white opacity-20"
+              style={{
+                left: `${Math.random() * 100}%`,
+                top: `${Math.random() * 100}%`,
+              }}
+              animate={{
+                y: [0, -100],
+                opacity: [0.2, 0],
+              }}
+              transition={{
+                duration: 10 + Math.random() * 10,
+                repeat: Number.POSITIVE_INFINITY,
+                delay: Math.random() * 10,
+              }}
+            />
+          ))}
+        </div>
+      </div>
+
+      <div className="container relative z-10 mx-auto px-4">
+        <div className="mx-auto max-w-4xl text-center">
+          <h2 className="mb-6 text-3xl font-bold tracking-tight md:text-4xl">
+            Ready to Transform Your Event Experience?
+          </h2>
+          <p className="mb-10 text-xl text-white/80">
+            Join thousands of event creators and participants who are already connecting through our platform.
+          </p>
+
+          <div className="flex flex-col justify-center gap-4 sm:flex-row">
+            <Button asChild size="lg" className="bg-white text-black hover:bg-gray-100">
+              <Link href="/dashboard/create-event">Create Event</Link>
+            </Button>
+
+            <Button asChild size="lg" variant="outline" className="border-white text-white hover:bg-white/10">
+              <Link href="/events">
+                Explore Events
+                <ArrowRight className="ml-2 h-4 w-4" />
+              </Link>
+            </Button>
+          </div>
+        </div>
+
+        {/* Stats */}
+        <div className="mx-auto mt-20 grid max-w-4xl gap-8 rounded-2xl border border-white/10 bg-white/5 p-8 backdrop-blur-sm sm:grid-cols-2 md:grid-cols-4">
+          {[
+            { value: "10K+", label: "Events Created" },
+            { value: "50K+", label: "Active Users" },
+            { value: "100+", label: "Countries" },
+            { value: "99%", label: "Satisfaction" },
+          ].map((stat, index) => (
+            <div key={index} className="text-center">
+              <p className="text-3xl font-bold text-white md:text-4xl">{stat.value}</p>
+              <p className="text-sm text-white/70 md:text-base">{stat.label}</p>
+            </div>
+          ))}
+        </div>
+      </div>
+    </section>
+  )
+}
